git-svn-id: http://xe-core.googlecode.com/svn/trunk@457 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ modules/board/board.controller.php
@@ -259,7 +259,7 @@
 
             // file class의 controller 객체 생성
             $oFileController = &getController('file');
-            $output = $oFileController->deleteFile($file_srl, $this->grant->manager);
+            if($file_srl) $output = $oFileController->deleteFile($file_srl, $this->grant->manager);
 
             // 첨부파일의 목록을 java script로 출력
             $oFileController->printUploadedFileList($document_srl);
